Remove some warnings (#97)
* Remove warnings in ACI

* Fix aci_build_F

* Fix asci

* Fix fci_mo

* Add mrci

* Add mrpt2

* Add determinant_hashvector and ci_reference

* Part of sparse_ci done

* Complete sparse_ci

* Fix ci_rdm

* Fix integrals

* Fix v2rdm

* Fix forte_old_methods and pymodule

* Fix test determinant

* Fix runtime error

* Update diskdf_integrals.cc

* Update fci_mo.cc

* Update fci_mo.cc

+++ src/sparse_ci/determinant_hashvector.cc
@@ -192,7 +192,7 @@ double DeterminantHashVec::overlap(SharedMatrix det1_evecs, int root1, Determina
 }
 
 void DeterminantHashVec::subspace(DeterminantHashVec& dets, SharedMatrix evecs,
-                                  std::vector<double>& new_evecs, int dim, int root) {
+                                  std::vector<double>& new_evecs, size_t dim, int root) {
     // Clear current wfn
     this->clear();
     new_evecs.assign(dim, 0.0);
